Moving Binding classes into android package.
