+++ formula-android/src/main/java/com/instacart/formula/android/Bindings.kt
@@ -1,4 +1,4 @@
-package com.instacart.formula.integration
+package com.instacart.formula.android
 
 class Bindings<in Component>(
     val types: Set<Class<*>>,
