+++ formula-android/src/main/java/com/instacart/formula/android/FeatureEvent.kt
@@ -1,7 +1,4 @@
-package com.instacart.formula.integration
-
-import com.instacart.formula.android.Feature
-import com.instacart.formula.android.FragmentId
+package com.instacart.formula.android
 
 sealed class FeatureEvent {
     data class Init(override val id: FragmentId, val feature: Feature<*>): FeatureEvent()
