+++ formula-android/src/main/java/com/instacart/formula/android/FeatureFactory.kt
@@ -3,7 +3,6 @@ package com.instacart.formula.android
 import com.instacart.formula.fragment.FormulaFragment
 import com.instacart.formula.fragment.FragmentKey
 import com.instacart.formula.fragment.FragmentFlowStore
-import com.instacart.formula.integration.FragmentBindingBuilder
 
 /**
  * Feature factory is responsible for creating a [Feature] for a specific [fragment key][Key].
