+++ formula-android/src/main/java/com/instacart/formula/android/FlowFactory.kt
@@ -1,7 +1,6 @@
 package com.instacart.formula.android
 
 import com.instacart.formula.integration.DisposableScope
-import com.instacart.formula.integration.FragmentBindingBuilder
 
 /**
  * A flow factory enables to group multiple fragments and share state, routers, action handlers
