+++ formula-android/src/main/java/com/instacart/formula/android/internal/CompositeBinding.kt
@@ -4,7 +4,7 @@ import com.instacart.formula.Evaluation
 import com.instacart.formula.Formula
 import com.instacart.formula.FormulaContext
 import com.instacart.formula.Stream
-import com.instacart.formula.integration.Binding
+import com.instacart.formula.android.Binding
 import com.instacart.formula.integration.DisposableScope
 
 /**
