+++ formula-android/src/main/java/com/instacart/formula/android/internal/FeatureBinding.kt
@@ -6,8 +6,8 @@ import com.instacart.formula.FormulaContext
 import com.instacart.formula.Stream
 import com.instacart.formula.android.FeatureFactory
 import com.instacart.formula.fragment.FragmentKey
-import com.instacart.formula.integration.Binding
-import com.instacart.formula.integration.FeatureEvent
+import com.instacart.formula.android.Binding
+import com.instacart.formula.android.FeatureEvent
 
 /**
  * Defines how a specific key should be bound to its [FeatureFactory],
