+++ formula-android/src/main/java/com/instacart/formula/android/internal/FeatureProvider.kt
@@ -1,6 +1,6 @@
 package com.instacart.formula.android.internal
 
-import com.instacart.formula.integration.FeatureEvent
+import com.instacart.formula.android.FeatureEvent
 import com.instacart.formula.android.FragmentId
 
 internal interface FeatureProvider {
