+++ formula-android/src/main/java/com/instacart/formula/fragment/FragmentFlowState.kt
@@ -1,7 +1,7 @@
 package com.instacart.formula.fragment
 
 import com.instacart.formula.android.FragmentId
-import com.instacart.formula.integration.FeatureEvent
+import com.instacart.formula.android.FeatureEvent
 import com.instacart.formula.android.FragmentState
 
 /**
